+++ max/python/max/pipelines/lib/diffusion_schedulers/__init__.py
@@ -1,0 +1,1 @@
+from .scheduler_factory import SchedulerFactory
